+++ src/main/java/ar/edu/unlam/tallerweb1/domain/usuarios/ServicioUsuario.java
@@ -11,5 +11,5 @@ public interface ServicioUsuario {
 
 	String encriptarClave(String clave) throws NoSuchAlgorithmException, UnsupportedEncodingException;
 
-	Usuario consultarUsuario(int id);
+	Usuario consultarUsuario(Long id);
 }
